+++ packages/lib-components/src/components/actions/IconButton/index.tsx
@@ -2,6 +2,7 @@ import { cloneElement, forwardRef, ReactElement } from 'react';
 import { Color, Size } from '../../../styled/types';
 import { Badge } from '../../../components';
 import { Default } from './style';
+import { getIconSize } from './getIconSize';
 
 export interface IconButtonProps {
   size?: Size;
@@ -13,21 +14,6 @@ export interface IconButtonProps {
   ariaLabel: string;
   badge?: string;
 }
-
-export const getIconSize = (size: Size) => {
-  switch (size) {
-    case 'tiny':
-      return 15;
-    case 'small':
-      return 16;
-    case 'medium':
-      return 18;
-    case 'large':
-      return 20;
-    case 'huge':
-      return 24;
-  }
-};
 
 export const IconButton = forwardRef<HTMLButtonElement, IconButtonProps>(function IconButton(
   { icon, color = 'primary', size = 'medium', disabled, onClick = () => {}, ariaLabel, badge },
