+++ packages/lib-components/src/components/data/Console/constants.ts
@@ -1,0 +1,6 @@
+export const DEFAULT_LINE_HEIGHT = 25;
+export const GUTTER_SIZE = 10;
+/*  TODO: this does not work with the scrollitemto because it has no knowledge of the padding.
+ *  https://gist.github.com/ValentinJS/4f3738ed426cac183e8446c869b162dd might be an option.
+ */
+export const LIST_PADDING_SIZE = 0;
