+++ packages/lib-components/src/components/data/Table/subcomponents/Filter/field.tsx
@@ -1,9 +1,10 @@
 import { Control, UseFormSetValue, useWatch } from 'react-hook-form';
 import { SelectField, TextField } from '../../../../../components';
-import { Operators, FilterInputType, IFormInputs } from '.';
+import { IFormInputs } from '.';
 import { Column } from '@tanstack/react-table';
 import { useLayoutEffect } from 'react';
 import { camelCaseToSpaces } from '../../../../../helpers';
+import { FilterInputType, Operators } from './types';
 
 interface FilterFieldProps<DataType> {
   index: number;
