+++ packages/lib-components/src/components/data/Table/subcomponents/Filter/types.ts
@@ -1,0 +1,10 @@
+export enum Operators {
+  is = 'is',
+  contains = 'contains',
+}
+
+export enum FilterInputType {
+  string = 'string',
+  uuid = 'uuid',
+  datetime = 'datetime',
+}
