+++ packages/lib-components/src/components/inputs/Date/types.ts
@@ -1,0 +1,14 @@
+export enum Tense {
+  Last = 'Last',
+  Next = 'Next',
+}
+
+export enum Unit {
+  Seconds = 'Seconds',
+  Minutes = 'Minutes',
+  Hours = 'Hours',
+  Days = 'Days',
+  Weeks = 'Weeks',
+  Months = 'Months',
+  Years = 'Years',
+}
