+++ packages/lib-components/src/components/inputs/TextField/Controlled.tsx
@@ -2,7 +2,7 @@ import { FC, useState } from 'react';
 import { useController } from 'react-hook-form';
 import { ControlledInputProps, defaultInputProps, defaultInputPropsFactory } from '../InputProps';
 import { TextFieldProps, GenericTextField } from '.';
-import { isNumber } from './Generic';
+import { isNumber } from './util';
 import { Container, InputContainer } from './style';
 import { Label, ErrorMessage, InputWrapper, Description } from '../layout';
