+++ packages/lib-components/src/components/inputs/layout/index.ts
@@ -1,6 +1,8 @@
-export { Description, setAriaDescribedBy } from './Description';
+export { Description } from './Description';
 export { InputWrapper } from './InputWrapper';
 export { ErrorMessage } from './ErrorMessage';
+
+export { setAriaDescribedBy } from './setAriaDescribedBy';
 
 export { Label } from './Label';
 export type { LabelProps, LabelPosition } from './Label';
