+++ packages/lib-components/src/components/inputs/layout/setAriaDescribedBy.ts
@@ -1,0 +1,3 @@
+export const setAriaDescribedBy = (name: string, hasDescription: boolean) => {
+  return hasDescription ? `${name}-description` : undefined;
+};
