+++ packages/lib-components/src/components/layout/index.ts
@@ -1,1 +1,1 @@
-export * from './Container';
+export { Container } from './Container';
