+++ packages/lib-components/src/components/navigation/index.ts
@@ -4,7 +4,8 @@ export type { StepperProps } from './Steppers/Stepper';
 export { SlimStepper } from './Steppers/SlimStepper';
 export type { SlimStepperProps } from './Steppers/SlimStepper';
 
-export { StepperProvider, useStepper } from './Steppers/context';
+export { StepperProvider } from './Steppers/provider';
+export { useStepper } from './Steppers/useStepper';
 
 export { IconNav } from './IconNav';
 export type { IconNavProps } from './IconNav';
