+++ packages/lib-components/src/components/other/PermissionsGuard/index.tsx
@@ -1,28 +1,12 @@
 import { FC, PropsWithChildren, ReactElement, useMemo } from 'react';
 import { PERMISSIONS } from '@takaro/apiclient';
-
-type PermissionsSet = PERMISSIONS | PERMISSIONS[];
-export type RequiredPermissions = PermissionsSet[];
+import { RequiredPermissions, hasPermissionHelper } from './hasPermissionsHelper';
 
 export interface PermissionsGuardProps {
   requiredPermissions: RequiredPermissions;
   userPermissions: PERMISSIONS[];
   fallback?: ReactElement;
 }
-
-export const hasPermissionHelper = (
-  userPermissions: PERMISSIONS[],
-  requiredPermissions: RequiredPermissions,
-): boolean => {
-  if (userPermissions.includes(PERMISSIONS.Root)) {
-    return true;
-  }
-  return requiredPermissions.some((permissionSet) =>
-    Array.isArray(permissionSet)
-      ? permissionSet.every((permission) => userPermissions.includes(permission))
-      : userPermissions.includes(permissionSet as PERMISSIONS),
-  );
-};
 
 export const PermissionsGuard: FC<PropsWithChildren<PermissionsGuardProps>> = ({
   userPermissions,
