+++ packages/lib-components/src/components/other/index.ts
@@ -12,8 +12,9 @@ export type { ActionMenuProps } from './ActionMenu';
 
 export { CollapseList } from './CollapseList';
 
-export { PermissionsGuard, hasPermissionHelper } from './PermissionsGuard';
-export type { PermissionsGuardProps, RequiredPermissions } from './PermissionsGuard';
+export { PermissionsGuard } from './PermissionsGuard';
+export { hasPermissionHelper, RequiredPermissions } from './PermissionsGuard/hasPermissionsHelper';
+export type { PermissionsGuardProps } from './PermissionsGuard';
 
 export { Collapsible } from './Collapsible';
 export type { CollapsibleProps } from './Collapsible';
