+++ packages/web-main/src/queries/item.tsx
@@ -4,7 +4,7 @@ import { AxiosError } from 'axios';
 import { getApiClient } from '../util/getApiClient';
 import { getNextPage, queryParamsToArray } from './util';
 
-export const itemKeys = {
+const itemKeys = {
   all: ['items'] as const,
   list: () => [...itemKeys.all] as const,
   detail: (id: string) => [...itemKeys.all, id] as const,
