+++ packages/web-main/src/routes/_auth/-module-builder/useModuleBuilderStore.tsx
@@ -1,12 +1,7 @@
-import {
-  CommandOutputDTO,
-  CronJobOutputDTO,
-  FunctionOutputDTO,
-  HookOutputDTO,
-  SmallModuleVersionOutputDTO,
-} from '@takaro/apiclient';
+import { SmallModuleVersionOutputDTO } from '@takaro/apiclient';
 import { PropsWithChildren, createContext, useContext, useRef } from 'react';
 import { createStore, useStore } from 'zustand';
+import { FileWithPath, FileMap } from './types';
 
 interface FileActions {
   openFile: (path: string) => void;
@@ -19,26 +14,6 @@ interface FileActions {
   closeFile: (path: string) => void;
   deleteFile: (path: string) => void;
 }
-
-export enum FileType {
-  Commands = 'commands',
-  Hooks = 'hooks',
-  CronJobs = 'cronjobs',
-  Functions = 'functions',
-}
-
-export type FileMap = Record<string, File>;
-
-interface File {
-  functionId: FunctionOutputDTO['id'];
-  type: FileType;
-  itemId: CommandOutputDTO['id'] | CronJobOutputDTO['id'] | HookOutputDTO['id'];
-  code: string;
-  hidden?: boolean;
-}
-
-// code is optional here
-type FileWithPath = Omit<File, 'code'> & { path: string; code?: string };
 
 export type ModuleBuilderProps = {
   moduleId: string;
