+++ packages/web-main/src/routes/_auth/_global/events.tsx
@@ -8,7 +8,8 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 import { hasPermission } from '../../../hooks/useHasPermission';
 import { useDocumentTitle } from '../../../hooks/useDocumentTitle';
 import { useMemo, useState } from 'react';
-import { EventFilter, EventFilterInputs, eventFilterSchema } from '../../../components/events/EventFilter';
+import { EventFilter, EventFilterInputs } from '../../../components/events/EventFilter';
+import { eventFilterSchema } from '../../../components/events/eventFilterSchema';
 import { PERMISSIONS } from '@takaro/apiclient';
 import { userMeQueryOptions } from '../../../queries/user';
