+++ packages/web-main/src/routes/_auth/gameserver.$gameServerId/settings.tsx
@@ -1,4 +1,3 @@
-import { booleanFields } from '../_global/settings/gameservers';
 import {
   useDeleteGameServerSetting,
   gameServerSettingsQueryOptions,
@@ -15,6 +14,7 @@ import { createFileRoute, redirect } from '@tanstack/react-router';
 import { useQueries } from '@tanstack/react-query';
 import { useDocumentTitle } from '../../../hooks/useDocumentTitle';
 import { userMeQueryOptions } from '../../../queries/user';
+import { booleanFields } from '../../../util/settings';
 
 export const Route = createFileRoute('/_auth/gameserver/$gameServerId/settings')({
   beforeLoad: async ({ context }) => {
