+++ packages/web-main/src/util/domainCookieRegex.ts
@@ -1,0 +1,1 @@
+export const TAKARO_DOMAIN_COOKIE_REGEX = /(?:(?:^|.*;\s*)takaro-domain\s*=\s*([^;]*).*$)|^.*$/;
