+++ packages/web-main/src/util/settings.ts
@@ -1,0 +1,2 @@
+// Boolean fields are strings in settings and need to be converted to boolean before being used.
+export const booleanFields = ['economyEnabled', 'developerMode'];
